Bumping GNU Unifont version to UnifontEX
So, the Rockbox (and the unofficial builds) use an ancient (2008) version of GNU Unifont, from long ago before emoji and such happened. This is bumping it to UnifontEX, which is based on Unifont 15.0.06-JP (the final and most complete TrueType build, from June 4th, 2023) plus Unifont 11.0.01 Upper (2018 Plane 1, this particular version is the highest possible version usable while remaining under 65535 glyphs. 11.0.02 Upper will fail merging with it, and it will ALSO fail merging with 11.0.01 base. So, we can get Plane 0 characters into 2023, and Plane 1 characters into 2018, which is better than having only characters from 2008.)